リクエスト本文の解釈の失敗を本物と同じ SerializationException と枠組みの検証にする
本物は型違いを SerializationException（AthenaErrorCode 無し）にし、文言は
JSON の型と目的の型の組み合わせで決まる。本文が JSON でなければ Message 無し、
必須の項目の欠落と null は枠組みの検証（Value null at '<lowerCamel>' ...
Member must not be null）、任意の null と未知のキーは無視、MaxResults の
Integer の範囲外は上限の検証になる（2026-09-23 に生 HTTP で 46 通りを実測）。
athena-local は serde の失敗を一律 InvalidRequestException と日本語の文言に
していたので、SDK が __type で選ぶ例外の型が本物と違っていた。

本文の解釈を request.rs に置き、JSON として読んでから null の項目を消し、
serde のエラー文（missing field / invalid type）で欠落と型違いを見分ける。
文言は実測した組み合わせだけ出し、未実測の組み合わせは Message 無しにする。
MaxResults は i64 で読み、範囲外を既存の上限の検査に流す。枠組みの検証の
文言の組み立ては response.rs に移して validation.rs と共有する。

+++ src/athena.rs
@@ -126,7 +126,7 @@ pub struct Statistics {
 pub struct GetQueryResultsRequest {
     pub query_execution_id: String,
     #[serde(default)]
-    pub max_results: Option<i32>,
+    pub max_results: Option<i64>,
     #[serde(default)]
     pub next_token: Option<String>,
 }
@@ -231,7 +231,7 @@ pub struct EngineVersion {
 #[serde(rename_all = "PascalCase")]
 pub struct ListWorkGroupsRequest {
     #[serde(default)]
-    pub max_results: Option<i32>,
+    pub max_results: Option<i64>,
     #[serde(default)]
     pub next_token: Option<String>,
 }
